remove logging, use nullable allof for pointers
A bug I found in my use of this so far is that depending on which model is registered first (whether its a pointer or not) determines whether an entire schema reference nullable. This means if some references to it are pointers, and some are not, we'll end up in a case where the result is nullable when it shouldn't be, or is non-nullable when it shouldn't be. Therefore, instead we should never make the reference model nullable, and instead when referencing it we can use the allof nullable combo to make it nullable if its a pointer. This may not be completely correct but I think it works for my needs